exception handling

+++ first-nestjs-project/src/tweet/tweet.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  RequestTimeoutException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Tweet } from './tweet.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -6,6 +10,7 @@ import { UsersService } from 'src/users/users.service';
 import { CreateTweetDto } from './dto/create.tweet.dto';
 import { HastagService } from 'src/hastag/hastag.service';
 import { UpdateTweetDto } from './dto/update.tweet.dto';
+import { ExceptionsHandler } from '@nestjs/core/exceptions/exceptions-handler';
 
 @Injectable()
 export class TweetService {
@@ -57,12 +62,6 @@ export class TweetService {
 
   //* ------------------- update tweet ------------------
   async updateTweet(updateTweet: UpdateTweetDto) {
-    // get hastags
-    updateTweet.hastags = updateTweet.hastags ?? []; // if the hastags comming from the user is null or undefined it will assign empty array
-    const hastags = await this.hastagService.getHastagArray(
-      updateTweet.hastags,
-    );
-
     // get tweet
     const tweet = await this.tweetRepository.findOneBy({ id: updateTweet.id });
     if (!tweet) {
@@ -71,27 +70,32 @@ export class TweetService {
       );
     }
     // creates objets for tweet
-    tweet.text = updateTweet.text ?? tweet.text;
-    tweet.image = updateTweet.image ?? tweet.image;
-    tweet.hastags = hastags;
+    if (updateTweet.text !== undefined) {
+      tweet.text = updateTweet.text;
+    }
+
+    if (updateTweet.image !== undefined) {
+      tweet.image = updateTweet.image;
+    }
+
+    if (updateTweet.hastags !== undefined) {
+      const hastags = await this.hastagService.getHastagArray(
+        updateTweet.hastags,
+      );
+      tweet.hastags = hastags;
+    }
 
     //save the changes
-    const respose = await this.tweetRepository.save(tweet);
-
-    // send response
-    return {
-      status: 'success',
-      message: 'Tweet updated successfully.',
-      data: { tweet: respose },
-    };
+    return await this.tweetRepository.save(tweet);
   }
 
   //* ------------------------delete tweet ---------------------
   async deleteTweet(id: number) {
-    await this.tweetRepository.delete({ id });
-    return {
-      stauts: 'success',
-      message: `A tweet with the id ${id} has been deleted!!`,
-    };
+    const tweet = this.tweetRepository.findOneBy({ id });
+    if (!tweet) {
+      throw new NotFoundException(`A tweet with the id:${id} doesn't exist.`);
+    }
+
+    return await this.tweetRepository.delete({ id });
   }
 }
